+++ src/ts/Ads/Ads.ts
@@ -271,6 +271,9 @@ export class Ads implements IAds {
         if (!this.isConsentShowRequired()) {
             return Promise.resolve();
         }
+
+        Diagnostics.trigger('consent_show', {adsConfig: JSON.stringify(this.Config.getDTO())});
+
         const consentView = new ConsentUnit({
             platform: this._core.NativeBridge.getPlatform(),
             privacyManager: this.PrivacyManager,
@@ -283,6 +286,10 @@ export class Ads implements IAds {
 
     public show(placementId: string, options: unknown, callback: INativeCallback): void {
         callback(CallbackStatus.OK);
+
+        if (!this._core.FocusManager.isAppForeground()) {
+            Diagnostics.trigger('ad_shown_in_background', {});
+        }
 
         if(this._showing) {
             // do not send finish event because there will be a finish event from currently open ad unit
@@ -465,12 +472,6 @@ export class Ads implements IAds {
             this._wasRealtimePlacement = false;
 
             this._currentAdUnit.show().then(() => {
-                if(this._core.NativeBridge.getPlatform() === Platform.ANDROID) {
-                    this._core.Api.Request.Android!.setMaximumPoolSize(8);
-                } else {
-                    this._core.Api.Request.setConcurrentRequestCount(8);
-                }
-
                 this.BackupCampaignManager.deleteBackupCampaigns();
             });
         });
@@ -491,12 +492,6 @@ export class Ads implements IAds {
 
     private onAdUnitClose(): void {
         this._showing = false;
-
-        if(this._core.NativeBridge.getPlatform() === Platform.ANDROID) {
-            this._core.Api.Request.Android!.setMaximumPoolSize(1);
-        } else {
-            this._core.Api.Request.setConcurrentRequestCount(1);
-        }
     }
 
     private setupTestEnvironment(): void {